loggen: add a log generator
It's compatible with logcnt in that it's producing a counter then random
fields.  The hostname is sent and defaults to the local host name though
it will be skipped if empty.

Right now a lot of cpu is spent in sendmsg(), probably because the kernel
needs to validate each and every component of the iovec. A better approach
would consist in memcpy() them into a temporary string that's change only
once in a while, and to keep only the header vectors and the counter. Maybe
even just make 3 vectors: header (changed once a second), the counter
(changed every send) and the data (changed once every N send).