loggen: replace the tiny random fields with a large lorem ipsum
The tiny fields were extremely expensive in kernel CPU time since an
access_ok() had to be performed for each of them. Let's change this for
a randomly large lorem ipsum (up to 5kB). The distribution is cubic so
that the size is centered around 1/8 of the max (i.e. ~660 bytes here).

The performance is now much better, with 1M logs/s being achievable at
100% CPU.