logcnt: implement a sliding window
On some multi-queue NICs, logs always arrive out-of-order. Let's
implement a small 64-entry queue made of one bit per packet to take
care of reordering. A packet is now considered as when the window
has to be advanced beyond 64+itself. In practice it now seems to
report accurate values that most of the time correspond to what is
really missing.