logcnt: don't count dups nor losses in the first window
During the first window we have a dilemma: if the first packet we receive
is not the first one, either we mark all predecessors as received so that
they're not counted as lost when the window shifts, but this might trigger
dups when the correct packets arrive, or we count them as not seen but they
then turn to losses.

What we're doing here is to consider as losses only what's received past
the first window. This allows us to avoid counting them as losses or dups,
so the report is more accurate.