loggen: use a freq_ctr to more accurately measure the rates
The generation rates (pkt and bit rates) were calculated as an average
from the startup, which left no margin for measurement errors, pauses
nor did it permit to implement slow rampups. Let's use freq_ctrs instead
to measure the pkt rate and the bit rate. Now the measurements are much
more accurate. In addition to this, it now allows to apply the two limits
in parallel, something that was not possible previously (pktrate would
take over). The bit rate limit is now in kbps though.