loggen: support starting multiple senders with -S
This creates as many sockets, host names and counters. These will appear
as multiple flows on the network by using several ports and will improve
the traffic distribution on multi-queue NICs. The limit is set to 1000
senders max, which should be sufficient and remains compatible with a
common fd limit of 1024.