loggen: preset the counters to the configured rate at boot
The problem with rate-limited values is that we start with blank counters
with no history, so we can send a very fast burst of initial packets up to
the configured packet rate, then pause, and only then start to smooth the
load.

Let's be smarter and pretend that the load was already at this level during
the previous period. This allows the rate-limiter to respect history and
maintain a constant load even during the first second. This will avoid a
huge number of losses that were always visible during startup.