loggen: also make the initial rampup smoother
The initial rampup can be jerky: usually the initial frequency
is extremely low, so the wait_us calculation is huge (typically one
second for an initial freq of 1). Because of this, short rampups cause
a one second delay, and then catch up with the speed on the next second,
with a zero history on the rate counter since it has expired in between,
causing a burst after a pause.

Here what we're doing is that we prevent from sleeping more than 1/10 of
the rampup time. That's sufficient to start soft and to progressively
catch up without this nasty effect.