logcnt: support counting only
This allows to count incoming messages without analysing them, either to
save a few bits of CPU cycles, or to accept invalid messages. On Linux
it enables usage of MSG_TRUNC which also avoids the memory copy, and it
uses recv() instead of recvfrom() so that we don't retrieve the sender's
address either. All of this combined allow to gain up to ~1-2% before
losses appear on some systems (particularly those dealing with kpti).